need long ints

+++ irish_townlands/management/commands/importosmhistory.py
@@ -10,7 +10,7 @@ from irish_townlands.models import Townland, ElectoralDivision, CivilParish, Bar
 logger = logging.getLogger(__name__)
 
 def clean_result_row(result):
-    result['osm_id'] = int(result['osm_id'])
+    result['osm_id'] = long(result['osm_id'])
     result['osm_timestamp'] = iso8601.parse_date(result['osm_timestamp'])
     result['osm_user'] = result['osm_user'].decode("utf8")
     return result
